+++ src/main/kotlin/apps/appCheckExaRegCurve/model/CheckRegCurvePost.kt
@@ -1,5 +1,5 @@
 package apps.appCheckExaRegCurve.model
 
-data class CheckRegCurvePost(val type : String, val annomeseIns : Long) {
+data class CheckRegCurvePost(val type : String, val annomeseIns : Int) {
     fun getJson() : String = JSON.stringify(this)
 }
