+++ src/main/kotlin/apps/appCheckExaRegCurve/model/InfoChartCheckRegCurve.kt
@@ -3,12 +3,12 @@ package apps.appCheckExaRegCurve.model
 
 
 class InfoChartCheckRegCurve(val type : String?,
-                             val minGGAnalisi : Long?,
-                             val maxGGAnalisi : Long?,
-                             val minAnnomeseIns : Long?,
-                             val maxAnnomeseIns : Long?,
+                             val minGGAnalisi : Int?,
+                             val maxGGAnalisi : Int?,
+                             val minAnnomeseIns : Int?,
+                             val maxAnnomeseIns : Int?,
                              val stati : Array<String>,
-                             val yearMonth : Array<Long>
+                             val yearMonth : Array<Int>
 ) {
     companion object {
         fun newEmptyRegCurveChartInfo()= InfoChartCheckRegCurve(null,null,null,null,null, emptyArray(),emptyArray())
